requirements.txt
Flask==3.0.3
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
cryptography==42.0.8
Werkzeug==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.31
